# api/swagger/swagger.yaml
swagger: "2.0"
info:
  version: "0.0.1"
  title: Movie Collection App
# during dev, should point to your local machine
host: localhost:10010
# basePath prefixes all resource paths 
basePath: /
# 
schemes:
  # tip: remove http to make production-grade
  - http
  - https
# format of bodies a client can send (Content-Type)
consumes:
  - application/json
# format of the responses to the client (Accepts)
produces:
  - application/json
paths:
  /movie:
    x-swagger-router-controller: movie
    get:
      operationId: getAllMovies
      description: Get the Movies list
      responses:
        "200":
          description: Success
          schema: 
            $ref: "#/definitions/GetMoviesListResponse"
        default:
          description: Error
          schema:
            $ref: "#/definitions/ErrorResponse"
    post:
      operationId: saveMovie
      description: Add a new Movie to the list
      parameters:
        - name: title
          description: Movie title
          in: body
          required: true
          schema:
            $ref: "#/definitions/Movie"
      responses:
        "200":
          description: Success
          schema:
            $ref: "#/definitions/GeneralResponse"
        default: 
          description: Error
          schema: 
            $ref: "#/definitions/ErrorResponse"     
  /movie/{id}:
    x-swagger-router-controller: movie
    get:
      operationId: getOneMovie
      description: Get a Movie
      parameters: 
        - name: id
          type: string
          in: path
          required: true
      responses:
        "200":
          description: Success
          schema:                                
            $ref: "#/definitions/GetMovieResponse"
        default: 
          description: Error
          schema:
            $ref: "#/definitions/ErrorResponse"     
    put:
      operationId: updateMovie
      description: Update a Movie
      parameters: 
        - name: id
          description: Movie ID
          type: string
          in: path
          required: true
        - name: title
          description: Movie title
          in: body
          required: true
          schema: 
            $ref: "#/definitions/Movie"
      responses: 
        "200":
          description: Success
          schema: 
            $ref: "#/definitions/GeneralResponse"       
        default: 
          description: Error
          schema:
            $ref: "#/definitions/ErrorResponse"    
    delete: 
      operationId: deleteMovie
      description: Delete a Movie
      parameters: 
        - name: id
          description: Movie ID        
          type: string
          in: path
          required: true
      responses:
        "200": 
          description: Success
          schema: 
            $ref: "#/definitions/GeneralResponse"
        default: 
          description: Error
          schema:
            $ref: "#/definitions/ErrorResponse"      

  /swagger:
    x-swagger-pipe: swagger_raw
definitions:
  GetMoviesListResponse:
    required:
      - movies
    properties:
      movies:
        type: array
        items:
          type: object
          properties:
            id:
              type: string
            title:
              type: string
            year:
              type: number  
  GetMovieResponse:
    required:
      - id            
      - title
      - year
    properties:
      id: 
        type: string
      title: 
        type: string
      year: 
        type: number      
  Movie:
    type: object
    properties:
      title:
        type: string
        description: Movie name
      year: 
        type: number
        description: Movie description 
    required:
      - title
      - year                           
  ErrorResponse:
    required:
      - message
    properties:
      message:
        type: string
  GeneralResponse: 
    type: object
    properties: 
      success: 
        type: number
        description: Returns 1 if successful
      description:
        type: string
        description: A short comment
    required:
      - success
      - description      